+++ src/Services/GoogleAds/ReportDownload.php
@@ -111,4 +111,15 @@ class ReportDownload
         return $report;
     }
 
+    /**
+     * toCollection()
+     *
+     *
+     * @return Collection results
+     */
+    public function toCollection()
+    {
+        return collect($this->toArray());
+    }
+
 }